scripts/requirements.txt
streamdiffusion[tensorrt]>=0.1.1
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
numpy>=1.24.0
